requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
cinemagoer==2023.5.1
colorama==0.4.6
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
IMDbPY==2022.7.9
jikanpy-v4==1.0.2
jproperties==2.1.1
lxml==4.9.3
multidict==6.0.4
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
requests==2.31.0
rottentomatoes-python==0.6.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
